refactor(types): removed duplicate types from tile.ts
