+++ next-tavla/src/Admin/scenarios/ColumnSettings/index.tsx
@@ -2,7 +2,8 @@ import React from 'react'
 import { useSortable } from '@dnd-kit/sortable'
 import { CSS, type Transform } from '@dnd-kit/utilities'
 import classNames from 'classnames'
-import { Columns, TColumn, TColumnSetting } from 'types/tile'
+import { Columns, TColumn } from 'types/column'
+import { TColumnSetting } from 'types/tile'
 import { DeleteIcon, AddIcon, SubtractIcon } from '@entur/icons'
 import { IconButton } from '@entur/button'
 import classes from './styles.module.css'
